Task #1
Task #1 is to create a bar chart or histogram to visualize the distribution of a categorical or continuous variable, such as the distribution of ages and genders in a population. 

I live in West Java province, Indonesia, so I decided to choose the Age and Gender distribution in West Java for 2021. I created a pie chart for the Gender Ratio in West Java for 2021 and a bar chart for the Age and Gender Distribution in West Java for 2021.